Refactor GraphQL queries and mutations
