+++ backend/src/redisPubSub.js
@@ -4,7 +4,7 @@ import IORedis from "ioredis";
 
 const options = {
   host: process.env.REDIS_HOST || "127.0.0.1",
-  port: process.env.REDIS_PORT || 6679,
+  port: process.env.REDIS_PORT || 6379,
   retryStrategy: (times) => Math.min(times * 50, 2000),
 };
