+++ frontend/src/graphql/aiProfileQueries.js
