--- frontend/src/graphql/matchMutations.js
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const DELETE_MATCH = gql`
-  mutation DeleteMatch($matchId: ID!) {
-    deleteMatch(matchId: $matchId)
-  }
-`;
