+++ frontend/src/graphql/photoQueries.js
