+++ frontend/src/graphql/swipeQueries.js
